+++ lib/dio_builder.dart
@@ -7,38 +7,21 @@ import 'package:dio_smart_retry/dio_smart_retry.dart';
 import 'package:pretty_dio_logger/pretty_dio_logger.dart';
 
 export 'dart:io';
-
 export 'package:chucker_flutter/chucker_flutter.dart';
 export 'package:dio/dio.dart';
 export 'package:dio_smart_retry/dio_smart_retry.dart';
 export 'package:pretty_dio_logger/pretty_dio_logger.dart';
 
+part 'dio_client.dart';
+
 /// The `DioBuilder` class provides a builder pattern for configuring and creating instances of the Dio HTTP client.
 /// It abstracts common configurations and interceptors for making HTTP requests.
-abstract class DioBuilder {
+class DioBuilder {
   static bool _allowBadRequest = false;
-  static bool _isLive = false;
-  static bool _isStage = false;
-  static bool _isTest = false;
-  static String _liveUrl = "";
-  static String _stageUrl = "";
-  static String _testUrl = "";
+  static String _baseApiUrl = ''; // Stores the base API URL
   bool _allowRetry = false;
 
-  final Dio _dio = Dio();
-
-  DioBuilder() {
-    addLogger();
-    allowRetryInFailed();
-    _dio.options = _buildBaseOption();
-    _dio.interceptors.add(_interceptorsWrapper);
-    if (_allowBadRequest) {
-      _dio.httpClientAdapter = Http2Adapter(ConnectionManager(
-        idleTimeout: const Duration(seconds: 10),
-        onClientCreate: (_, config) => config.onBadCertificate = (_) => true,
-      ));
-    }
-  }
+  static final Dio _dio = Dio();
 
   ResponseType _responseType = ResponseType.json;
   String _contentType = "application/json";
@@ -52,6 +35,28 @@ abstract class DioBuilder {
   Map<String, dynamic>? _queryParameters;
   bool _receiveDataWhenStatusError = true;
   int _defaultTimeOut = 15;
+
+  /// Initializes the DioBuilder instance with default configurations.
+  void initialize() {
+    _clearInterceptors();
+    addLogger();
+    allowRetryInFailed();
+    _dio.options = _buildBaseOption();
+    _dio.interceptors.add(_interceptorsWrapper);
+    if (_allowBadRequest) {
+      _dio.httpClientAdapter = Http2Adapter(ConnectionManager(
+        idleTimeout: const Duration(seconds: 10),
+        onClientCreate: (_, config) => config.onBadCertificate = (_) => true,
+      ));
+    }
+  }
+
+  /// Clears old interceptors if they exist, called every time you call initialize.
+  void _clearInterceptors() {
+    if (_dio.interceptors.isNotEmpty) {
+      _dio.interceptors.clear();
+    }
+  }
 
   /// Adds a logger interceptor to the Dio client for logging HTTP requests and responses.
   /// The logger can be printed to the console or integrated with the Chucker library for more detailed logs.
@@ -89,24 +94,6 @@ abstract class DioBuilder {
     return this;
   }
 
-  /// Sets the base URL for the live environment.
-  /// This URL is used when the application is in production and deployed to app stores.
-  void setBaseLiveUrl(String baseUrl) {
-    _liveUrl = baseUrl;
-  }
-
-  /// Sets the base URL for the testing or development environment.
-  DioBuilder setBaseTestUrl(String baseUrl) {
-    _testUrl = baseUrl;
-    return this;
-  }
-
-  /// Sets the base URL for the staging environment.
-  DioBuilder setBaseStageUrl(String baseUrl) {
-    _stageUrl = baseUrl;
-    return this;
-  }
-
   /// Enables or disables the ability to handle bad SSL certificates or invalid HTTPS connections.
   /// This can be useful for testing or development purposes but should be used with caution.
   DioBuilder allowBadRequest(bool value) {
@@ -114,75 +101,77 @@ abstract class DioBuilder {
     return this;
   }
 
-  /// Sets the environment to live mode.
-  DioBuilder setIsLive(bool value) {
-    _isLive = value;
-    return this;
-  }
-
-  /// Sets the environment to test mode.
-  DioBuilder setIsTest(bool value) {
-    _isTest = value;
-    return this;
-  }
-
-  /// Sets the environment to stage mode.
-  DioBuilder setIsStage(bool value) {
-    _isStage = value;
-    return this;
-  }
-
   /// Sets the default timeout for network requests, in seconds.
   DioBuilder setDefaultTimeOut(int value) {
     _defaultTimeOut = value;
+    _dio.options.connectTimeout = Duration(seconds: value);
+    _dio.options.receiveTimeout = Duration(seconds: value);
+    _dio.options.sendTimeout = Duration(seconds: value);
+    initialize();
     return this;
   }
 
   /// Sets the response type for the HTTP requests.
   DioBuilder setResponseType(ResponseType value) {
     _responseType = value;
+    _dio.options.responseType = value;
+    initialize();
     return this;
   }
 
   /// Sets the content type for the HTTP requests.
   DioBuilder setContentType(String value) {
     _contentType = value;
+    _dio.options.contentType = value;
+    initialize();
     return this;
   }
 
   /// Sets whether the connection should be persistent or not.
   DioBuilder setPersistentConnection(bool value) {
     _persistentConnection = value;
+    _dio.options.persistentConnection = value;
+    initialize();
     return this;
   }
 
   /// Sets whether the client should follow redirects or not.
   DioBuilder setFollowRedirects(bool value) {
     _followRedirects = value;
+    _dio.options.followRedirects = value;
+    initialize();
     return this;
   }
 
   /// Sets the maximum number of redirects to follow.
   DioBuilder setMaxRedirects(int value) {
     _maxRedirects = value;
+    _dio.options.maxRedirects = value;
+    initialize();
     return this;
   }
 
   /// Sets extra options to be included in the requests.
   DioBuilder setExtras(Map<String, dynamic> value) {
     _extra = value;
+    _dio.options.extra = value;
+    initialize();
     return this;
   }
 
   /// Sets the headers to be included in the requests.
   DioBuilder setHeaders(Map<String, dynamic> value) {
     _headers = value;
+    _dio.options.headers = _headers;
+    initialize();
     return this;
   }
 
   /// Sets the query parameters to be included in the requests.
   DioBuilder queryParameters(Map<String, dynamic> value) {
     _queryParameters = value;
+    _dio.options.queryParameters = value;
+    initialize();
     return this;
   }
 
@@ -194,12 +183,16 @@ abstract class DioBuilder {
   /// Sets the format for lists serialization in the requests.
   DioBuilder setListFormat(ListFormat value) {
     _listFormat = value;
+    _dio.options.listFormat = value;
+    initialize();
     return this;
   }
 
   /// Sets the HTTP method for the requests.
   DioBuilder setMethod(String value) {
     _method = value;
+    _dio.options.method = value;
+    initialize();
     return this;
   }
 
@@ -210,9 +203,9 @@ abstract class DioBuilder {
 
   /// Checks if there is an internet connection between the client and server.
   /// Returns `true` if there is a connection, otherwise `false`.
-  Future<bool> checkInternetConnection() async {
+  Future<bool> checkInternetConnection({String? lookUpAddress}) async {
     try {
-      final result = await InternetAddress.lookup(_baseUrl());
+      final result = await InternetAddress.lookup(lookUpAddress ?? baseUrl);
       return result.isNotEmpty && result[0].rawAddress.isNotEmpty;
     } on SocketException {
       return false;
@@ -226,7 +219,7 @@ abstract class DioBuilder {
 
   /// Builds the base options for the Dio client.
   BaseOptions _buildBaseOption() => BaseOptions(
-        baseUrl: _baseUrl(),
+        baseUrl: baseUrl,
         connectTimeout: Duration(seconds: _defaultTimeOut),
         receiveTimeout: Duration(seconds: _defaultTimeOut),
         sendTimeout: Duration(seconds: _defaultTimeOut),
@@ -244,17 +237,13 @@ abstract class DioBuilder {
       );
 
   /// Returns the base URL based on the current environment.
-  String _baseUrl() {
-    if (_isLive && _liveUrl.isNotEmpty) {
-      return _liveUrl;
-    } else if (_isStage && _stageUrl.isNotEmpty) {
-      return _stageUrl;
-    } else if (_isTest && _testUrl.isNotEmpty) {
-      return _testUrl;
-    } else {
-      throw Exception("No URL has been allowed for use yet.");
-    }
+  set baseUrl(String baseUrl) {
+    _baseApiUrl = baseUrl;
   }
+
+  String get baseUrl => _baseApiUrl.isNotEmpty
+      ? _baseApiUrl
+      : throw Exception("No URL has been set yet.");
 
   /// Wraps the interceptors for error, request, and response.
   InterceptorsWrapper get _interceptorsWrapper => InterceptorsWrapper(
@@ -266,13 +255,19 @@ abstract class DioBuilder {
 
   /// Handles the error that occurred during the HTTP request.
   void handleOnError(
-      DioException dioException, ErrorInterceptorHandler handler);
+      DioException dioException, ErrorInterceptorHandler handler){
+    handler.next(dioException);
+  }
 
   /// Handles the request before it is sent.
   void handleOnRequest(
-      RequestOptions options, RequestInterceptorHandler handler);
+      RequestOptions options, RequestInterceptorHandler handler){
+    handler.next(options);
+  }
 
   /// Handles the response after it is received.
   void handleOnResponse(
-      Response<dynamic> response, ResponseInterceptorHandler handler);
+      Response<dynamic> response, ResponseInterceptorHandler handler){
+    handler.next(response);
+  }
 }